fix(empty-room-baseline): register module + 4 correctness fixes
Addresses 5 PR-458 review threads:

  1. (codex P1) ALSO catches a PR 3 regression. Adding the module to
     build_src_filter just compiles it; csi_module_register() makes
     it tick. The canary runtime's securacv_csi_modules_init only
     registered presence/breathing/activity/daily/anomaly — both PR 3's
     core.multilink_fusion AND this PR's meta.empty_room_baseline were
     dead code on device. Fixed in
     firmware/canary/src/csi_modules_integration.cpp.

  2. (gemini MED) s_in_progress_n was uint16_t; overflowed at ~18.2 h
     at 1 Hz windows. API accepts duration_ms up to UINT32_MAX (~49
     days). Promoted to uint32_t. s_baseline_n same. csi_event_values_t.
     bundled_count stays uint16_t (chokepoint contract) — capped at
     emit time.

  3. (gemini MED) on_tick checked the deadline AFTER the
     `if (f == nullptr) return` guard, so a stalled CSI stream left
     the module stuck in CALIBRATING. Moved the deadline check first.
     Used signed-delta `(int32_t)(now - (started + duration)) >= 0`
     for wrap-safety across the uint32_t millis() rollover.

  4. (gemini MED) Event type_name was `baseline_calibrated` but
     carried state_name = "calibrated" | "failed" | "cancelled".
     Renamed to `baseline_status`. Consumers filter by state_name.

  5. (gemini HIGH) Thread-safety: added a docstring stating the
     single-task contract — start/cancel/get + on_tick all run on the
     csi_module dispatcher's task (matches the existing csi_module
     spec: "tick() is called from the main loop, NOT an ISR"). No
     atomics added; the cross-task path doesn't exist in production
     today and silent locks would hide future violations rather than
     surface them.

Plus a test-harness fix discovered during the verification: fresh_world()
called meta_empty_room_baseline_test_reset() AFTER setting the virtual
clock, which wiped the clock back to its default fake counter. With the
new deadline-first ordering in on_tick that surfaced as a finalize-too-
early bug. Reordered fresh_world to set the clock LAST.

6/6 host tests still pass. check_csi_sync mirror updated. Regression +
all existing suites green.

+++ firmware/common/csi/src/meta_empty_room_baseline.h
@@ -99,7 +99,7 @@ void meta_empty_room_baseline_test_reset(void);
 void meta_empty_room_baseline_test_set_now_ms(uint32_t now_ms);
 
 /** Inspect the accumulator: how many windows have been added so far. */
-uint16_t meta_empty_room_baseline_test_in_progress_count(void);
+uint32_t meta_empty_room_baseline_test_in_progress_count(void);
 #endif
 
 #ifdef __cplusplus
